Add client registrar
