+++ src/main/java/com/zebrunner/agent/core/registrar/ClientRegistrar.java
@@ -1,0 +1,18 @@
+package com.zebrunner.agent.core.registrar;
+
+public class ClientRegistrar {
+
+    private static volatile ZebrunnerApiClient client;
+
+    public static synchronized ZebrunnerApiClient getClient() {
+        if (client == null) {
+            return DefaultZebrunnerApiClient.getInstance();
+        }
+        return client;
+    }
+
+    public static synchronized void register(ZebrunnerApiClient newClient) {
+        client = newClient;
+    }
+
+}
