+++ src/main/java/com/zebrunner/agent/core/registrar/CurrentTest.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class CurrentTest {
 
-    private static final ZebrunnerApiClient API_CLIENT = ZebrunnerApiClient.getInstance();
+    private static final ZebrunnerApiClient API_CLIENT = ClientRegistrar.getClient();
 
     /**
      * This method returns Zebrunner Test id.
